Upstream version 0.17.1

+++ src/tube-stream.h
@@ -22,9 +22,8 @@
 
 #include <glib-object.h>
 
-#include <telepathy-glib/enums.h>
-#include <telepathy-glib/interfaces.h>
-#include <telepathy-glib/base-channel.h>
+#include <telepathy-glib/telepathy-glib.h>
+#include <telepathy-glib/telepathy-glib-dbus.h>
 
 #include "connection.h"
 #include "extensions/extensions.h"
@@ -69,7 +68,7 @@ GType gabble_tube_stream_get_type (void);
 GabbleTubeStream *gabble_tube_stream_new (GabbleConnection *conn,
     TpHandle handle, TpHandleType handle_type, TpHandle self_handle,
     TpHandle initiator, const gchar *service, GHashTable *parameters,
-    guint id, GabbleMucChannel *muc, gboolean requested);
+    guint64 id, GabbleMucChannel *muc, gboolean requested);
 
 gboolean gabble_tube_stream_check_params (TpSocketAddressType address_type,
     const GValue *address, TpSocketAccessControl access_control,
